Introduce OkHttp3 test and properties file
This commit adds a dependency to OkHttp3 and a simple test class to
play around with it a bit. At the moment it sends the Contact (aka Handle
) List enquiry, created via the corresponding builder, via OkHttp to the
AutoDNS demo server.
The authentication for this is in a properties file which needs to be in
the root directory and must have the name "autodns.properties". An
example of this file is also part of this commit
(`autodns.sample.properties`) and a small addition to `.gitignore` should
pretend you from commiting your production / real life logins to GitHub.