Ignoring log config and NIC reference list
The easy thing first: A new `.gitignore` will find its way into the
example sub project. It will ignore my log configuration.

More interesting is the new enum `NicReferences`. After a lot of digging
I was able to extract a list of "valid" NIC references - needed when a
contact is supposed to be able to create certain domains (or act as any
kind of contact for this matter).
The documentation is rather bad in this respect, as it's defining the
attribute `nic_ref` of type `country`, which is definitely wrong.
`NicReferences` will contain a list of selectable values from the AutoDNS
UI (this is correct as of 28/10/2017). So it needs verification.
In a later change the new enum will get an adapter and will be the
correct type for the `nic_ref` list.